refactor(agents): split auth service, extract overlay policy and login
auth/service.rs owns AgentAuthService (apply/status/launch_overlay as
resolve -> decide -> materialize); auth/overlay_policy.rs is the pure
planner; auth/login.rs takes the login-command resolution out of the
runtime facade; runtime.rs slims to sentence-shaped use-cases with one
instrument span on install.
