+++ anyharness/crates/anyharness-lib/src/domains/agents/auth/claude_tests.rs
@@ -5,7 +5,7 @@ use anyharness_contract::v1::AgentAuthSelectionConfig;
 use crate::domains::sessions::mcp_bindings::crypto::SessionDataCipher;
 use crate::persistence::Db;
 
-use super::{AgentAuthConfigInput, AgentAuthConfigService, AgentAuthConfigStore};
+use super::{AgentAuthConfigInput, AgentAuthService, AgentAuthConfigStore};
 
 fn cipher() -> SessionDataCipher {
     SessionDataCipher::from_env_value("MDEyMzQ1Njc4OWFiY2RlZjAxMjM0NTY3ODlhYmNkZWY=")
@@ -19,7 +19,7 @@ fn claude_gateway_launch_overlay_sets_managed_config_dir() {
         std::process::id()
     ));
     let _ = std::fs::remove_dir_all(&root);
-    let service = AgentAuthConfigService::new(
+    let service = AgentAuthService::new(
         AgentAuthConfigStore::new(Db::open_in_memory().expect("db")),
         Some(cipher()),
         root.clone(),
@@ -73,7 +73,7 @@ fn claude_gateway_launch_overlay_sets_managed_config_dir() {
 
 #[test]
 fn apply_config_rejects_support_env_claude_config_dir() {
-    let service = AgentAuthConfigService::new(
+    let service = AgentAuthService::new(
         AgentAuthConfigStore::new(Db::open_in_memory().expect("db")),
         Some(cipher()),
         std::env::temp_dir(),
